Extract shared styles and components to eliminate duplication
- Add SHARED_STYLES to theme.ts for common overlay, button, and layout patterns
- Create LoadingScreen component replacing 4 duplicate implementations
- Create CreditsContent component eliminating duplicate credits code
- Update all game screens to use shared styles
