+++ mobile/game/components/free-play-menu.tsx
@@ -8,11 +8,11 @@ import {
   StyleSheet,
   Text,
   Pressable,
-  ActivityIndicator,
   Alert,
 } from 'react-native';
 import { SafeAreaWrapper } from '@/components/safe-area-wrapper';
-import { GAME_COLORS, FONT_SIZES, SPACING } from '@/constants/theme';
+import { LoadingScreen } from '@/components/loading-screen';
+import { GAME_COLORS, FONT_SIZES, SPACING, SHARED_STYLES } from '@/constants/theme';
 import {
   FREE_PLAY_SLOTS,
   type FreePlaySlot,
@@ -150,11 +150,7 @@ export function FreePlayMenu({ onSelectSlot, onBack }: FreePlayMenuProps) {
   );
 
   if (isLoading) {
-    return (
-      <View style={styles.loadingContainer}>
-        <ActivityIndicator size="large" color={GAME_COLORS.TEXT_PRIMARY} />
-      </View>
-    );
+    return <LoadingScreen />;
   }
 
   return (
@@ -184,22 +180,12 @@ export function FreePlayMenu({ onSelectSlot, onBack }: FreePlayMenuProps) {
 
 const styles = StyleSheet.create({
   container: {
-    flex: 1,
-    backgroundColor: '#0a1015',
-    padding: SPACING.LARGE.INT,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  loadingContainer: {
-    flex: 1,
-    backgroundColor: '#0a1015',
+    ...SHARED_STYLES.screenContainer,
     justifyContent: 'center',
     alignItems: 'center',
   },
   title: {
-    fontFamily: 'DepartureMonoRegular',
-    fontSize: FONT_SIZES.HUGE.INT,
-    color: GAME_COLORS.TEXT_PRIMARY,
+    ...SHARED_STYLES.titleText,
     marginBottom: SPACING.SMALL.INT,
   },
   subtitle: {
@@ -248,17 +234,8 @@ const styles = StyleSheet.create({
     marginTop: SPACING.LARGE.INT,
     marginBottom: SPACING.XLARGE.INT,
   },
-  backButton: {
-    backgroundColor: GAME_COLORS.BUTTON_PRIMARY,
-    paddingVertical: SPACING.SMALL.INT,
-    paddingHorizontal: SPACING.XLARGE.INT,
-    borderRadius: 4,
-  },
-  backButtonText: {
-    fontFamily: 'DepartureMonoRegular',
-    fontSize: FONT_SIZES.MEDIUM.INT,
-    color: GAME_COLORS.TEXT_PRIMARY,
-  },
+  backButton: SHARED_STYLES.backButton,
+  backButtonText: SHARED_STYLES.buttonText,
 });
 
 export default FreePlayMenu;
