+++ mobile/game/components/puzzle-world-select.tsx
@@ -9,10 +9,10 @@ import {
   Text,
   Pressable,
   ScrollView,
-  ActivityIndicator,
 } from 'react-native';
 import { SafeAreaWrapper } from '@/components/safe-area-wrapper';
-import { GAME_COLORS, FONT_SIZES, SPACING } from '@/constants/theme';
+import { LoadingScreen } from '@/components/loading-screen';
+import { GAME_COLORS, FONT_SIZES, SPACING, SHARED_STYLES } from '@/constants/theme';
 import { getWorldsMetadata, type PuzzleProgress } from '../modes/puzzle';
 import { loadPuzzleProgress } from '@/services/storage';
 
@@ -107,11 +107,7 @@ export function PuzzleWorldSelect({ onSelectWorld, onBack }: PuzzleWorldSelectPr
   );
 
   if (isLoading) {
-    return (
-      <View style={styles.loadingContainer}>
-        <ActivityIndicator size="large" color={GAME_COLORS.TEXT_PRIMARY} />
-      </View>
-    );
+    return <LoadingScreen />;
   }
 
   return (
@@ -146,21 +142,10 @@ export function PuzzleWorldSelect({ onSelectWorld, onBack }: PuzzleWorldSelectPr
 
 const styles = StyleSheet.create({
   container: {
-    flex: 1,
-    backgroundColor: '#0a1015',
-    padding: SPACING.LARGE.INT,
-  },
-  loadingContainer: {
-    flex: 1,
-    backgroundColor: '#0a1015',
-    justifyContent: 'center',
-    alignItems: 'center',
+    ...SHARED_STYLES.screenContainer,
   },
   title: {
-    fontFamily: 'DepartureMonoRegular',
-    fontSize: FONT_SIZES.HUGE.INT,
-    color: GAME_COLORS.TEXT_PRIMARY,
-    textAlign: 'center',
+    ...SHARED_STYLES.titleText,
     marginBottom: SPACING.SMALL.INT,
   },
   subtitle: {
@@ -219,18 +204,10 @@ const styles = StyleSheet.create({
     color: GAME_COLORS.TEXT_SECONDARY,
   },
   backButton: {
-    backgroundColor: GAME_COLORS.BUTTON_PRIMARY,
-    paddingVertical: SPACING.SMALL.INT,
-    paddingHorizontal: SPACING.XLARGE.INT,
-    borderRadius: 4,
-    alignSelf: 'center',
+    ...SHARED_STYLES.backButton,
     marginTop: SPACING.MEDIUM.INT,
   },
-  backButtonText: {
-    fontFamily: 'DepartureMonoRegular',
-    fontSize: FONT_SIZES.MEDIUM.INT,
-    color: GAME_COLORS.TEXT_PRIMARY,
-  },
+  backButtonText: SHARED_STYLES.buttonText,
 });
 
 export default PuzzleWorldSelect;
